separate the function of token-generation
